chore: upgrade dependencies and improve code quality - Upgrade Go version in CI workflow to 1.25 and 1.26 - Update actions/checkout and actions/setup-go to v6 - Upgrade codecov-action to v6 - Refactor various functions to use context for better cancellation support - Improve variable naming for clarity - Add verification step in Makefile for Go installation - Update linters and their settings in .golangci.yml

+++ gitlab_release.go
@@ -76,7 +76,7 @@ type GitLabAsset struct {
 
 func NewGitLabAsset(from *gitlab.ReleaseLink) *GitLabAsset {
 	return &GitLabAsset{
-		id:   int64(from.ID),
+		id:   from.ID,
 		name: from.Name,
 		url:  from.URL,
 	}
